/reset-password route added
