+++ src/modules/account/components/login/index.tsx
@@ -5,6 +5,7 @@ import Input from "@modules/common/components/input"
 import { logCustomerIn } from "@modules/account/actions"
 import ErrorMessage from "@modules/checkout/components/error-message"
 import { SubmitButton } from "@modules/checkout/components/submit-button"
+import { useRouter } from "next/navigation"
 
 type Props = {
   setCurrentView: (view: LOGIN_VIEW) => void
@@ -12,6 +13,11 @@ type Props = {
 
 const Login = ({ setCurrentView }: Props) => {
   const [message, formAction] = useFormState(logCustomerIn, null)
+  const router = useRouter()
+
+  const handleResetPassword = () => {
+    router.push("/reset-password") // Navigate to the reset-password page
+  }
 
   return (
     <div className="max-w-sm w-full flex flex-col items-center" data-testid="login-page">
@@ -45,7 +51,7 @@ const Login = ({ setCurrentView }: Props) => {
       <span className="text-center text-ui-fg-base text-small-regular mt-6">
         Forgot your password{" "}
         <button
-          onClick={() => setCurrentView(LOGIN_VIEW.RESET_PASSWORD)}
+          onClick={handleResetPassword}
           className="underline"
         >
           Reset Password
