+++ src/modules/account/templates/login-template.tsx
@@ -4,12 +4,11 @@ import { useState } from "react"
 
 import Register from "@modules/account/components/register"
 import Login from "@modules/account/components/login"
-import ResetPassword from "../components/reset-password"
+
 
 export enum LOGIN_VIEW {
   SIGN_IN = "sign-in",
   REGISTER = "register",
-  RESET_PASSWORD = "reset-password",
   CHOOSE_NEW_PASSWORD = "choose-new-password"
 }
 
@@ -20,8 +19,6 @@ const LoginTemplate = () => {
     <div className="w-full flex justify-center px-8 py-8">
     {currentView === "sign-in" ? (
       <Login setCurrentView={setCurrentView} />
-    ) : currentView === "reset-password" ? (
-      <ResetPassword setCurrentView={setCurrentView} />
     ) : currentView === "register" ? (
       <Register setCurrentView={setCurrentView} />
     ) : null}
